# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NoStress Events API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EventStatus = (typeof EventStatus)[keyof typeof EventStatus];

export const EventStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface TicketType {
  id: string;
  name: string;
  nameFr?: string;
  price: number;
  currency?: string;
  available?: number;
  description?: string;
}

export interface Event {
  id: string;
  title: string;
  titleFr?: string;
  description?: string;
  descriptionFr?: string;
  date: string;
  time?: string;
  venue: string;
  venueId?: string;
  city: string;
  category: string;
  imageUrl?: string;
  price: number;
  currency?: string;
  isSponsored?: boolean;
  status: EventStatus;
  ticketTypes?: TicketType[];
  latitude?: number;
  longitude?: number;
  createdAt?: string;
}

export interface EventsListResponse {
  events: Event[];
  total: number;
  page: number;
  limit: number;
}

export interface CreateEventRequest {
  title: string;
  titleFr?: string;
  description?: string;
  descriptionFr?: string;
  date: string;
  time?: string;
  venue: string;
  venueId?: string;
  city: string;
  category: string;
  imageUrl?: string;
  price: number;
  currency?: string;
  isSponsored?: boolean;
  ticketTypes?: TicketType[];
  latitude?: number;
  longitude?: number;
}

export type UpdateEventRequestStatus =
  (typeof UpdateEventRequestStatus)[keyof typeof UpdateEventRequestStatus];

export const UpdateEventRequestStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface UpdateEventRequest {
  title?: string;
  titleFr?: string;
  description?: string;
  descriptionFr?: string;
  date?: string;
  time?: string;
  venue?: string;
  venueId?: string;
  city?: string;
  category?: string;
  imageUrl?: string;
  price?: number;
  currency?: string;
  isSponsored?: boolean;
  status?: UpdateEventRequestStatus;
  ticketTypes?: TicketType[];
}

export interface Venue {
  id: string;
  name: string;
  type: string;
  city: string;
  address?: string;
  imageUrl?: string;
  description?: string;
  isVerified?: boolean;
  latitude?: number;
  longitude?: number;
  ownerId?: string;
  createdAt?: string;
}

export interface VenuesListResponse {
  venues: Venue[];
  total: number;
}

export interface CreateVenueRequest {
  name: string;
  type: string;
  city: string;
  address?: string;
  imageUrl?: string;
  description?: string;
  latitude?: number;
  longitude?: number;
}

export interface UpdateVenueRequest {
  name?: string;
  type?: string;
  city?: string;
  address?: string;
  imageUrl?: string;
  description?: string;
  latitude?: number;
  longitude?: number;
}

export interface City {
  id: string;
  name: string;
  country: string;
  imageUrl?: string;
}

export interface CitiesListResponse {
  cities: City[];
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  email: string;
  password: string;
  name: string;
  phone?: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  structure: "structure",
  admin: "admin",
} as const;

export interface User {
  id: string;
  email: string;
  name: string;
  phone?: string;
  role: UserRole;
  avatarUrl?: string;
  favorites?: string[];
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type PurchaseTicketRequestPaymentMethod =
  (typeof PurchaseTicketRequestPaymentMethod)[keyof typeof PurchaseTicketRequestPaymentMethod];

export const PurchaseTicketRequestPaymentMethod = {
  flooz: "flooz",
  tmoney: "tmoney",
  mix_yas: "mix_yas",
} as const;

export interface PurchaseTicketRequest {
  eventId: string;
  ticketTypeId: string;
  quantity: number;
  paymentMethod: PurchaseTicketRequestPaymentMethod;
  phoneNumber: string;
}

export type TicketPurchaseResponseTicket = {
  id?: string;
  eventId?: string;
  ticketType?: string;
  qrCode?: string;
};

export interface TicketPurchaseResponse {
  transactionId: string;
  status: string;
  ticket?: TicketPurchaseResponseTicket;
}

export interface SubscriptionPlan {
  id: string;
  name: string;
  nameFr?: string;
  monthlyPriceFCFA: number;
  features: string[];
  featuresFr?: string[];
  isPopular?: boolean;
}

export interface SubscriptionPlansResponse {
  plans: SubscriptionPlan[];
}

export type GetEventsParams = {
  city?: string;
  category?: string;
  page?: number;
  limit?: number;
};

export type GetVenuesParams = {
  city?: string;
  type?: string;
};