# lib/api-zod/src/generated/types/citiesListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NoStress Events API
 * OpenAPI spec version: 0.1.0
 */
import type { City } from "./city";

export interface CitiesListResponse {
  cities: City[];
}