# lib/api-zod/src/generated/types/eventsListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NoStress Events API
 * OpenAPI spec version: 0.1.0
 */
import type { Event } from "./event";

export interface EventsListResponse {
  events: Event[];
  total: number;
  page: number;
  limit: number;
}