# lib/api-zod/src/generated/types/subscriptionPlansResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NoStress Events API
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriptionPlan } from "./subscriptionPlan";

export interface SubscriptionPlansResponse {
  plans: SubscriptionPlan[];
}