# lib/api-zod/src/generated/types/ticketPurchaseResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NoStress Events API
 * OpenAPI spec version: 0.1.0
 */
import type { TicketPurchaseResponseTicket } from "./ticketPurchaseResponseTicket";

export interface TicketPurchaseResponse {
  transactionId: string;
  status: string;
  ticket?: TicketPurchaseResponseTicket;
}