# lib/api-zod/src/generated/types/ticketType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NoStress Events API
 * OpenAPI spec version: 0.1.0
 */

export interface TicketType {
  id: string;
  name: string;
  nameFr?: string;
  price: number;
  currency?: string;
  available?: number;
  description?: string;
}