# lib/api-zod/src/generated/types/venue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NoStress Events API
 * OpenAPI spec version: 0.1.0
 */

export interface Venue {
  id: string;
  name: string;
  type: string;
  city: string;
  address?: string;
  imageUrl?: string;
  description?: string;
  isVerified?: boolean;
  latitude?: number;
  longitude?: number;
  ownerId?: string;
  createdAt?: string;
}